+++ packages/step/step.ts
@@ -1,18 +1,19 @@
 import { classNames } from '@chbphone55/classnames';
 import { i18n } from '@lingui/core';
-import { html, LitElement } from 'lit';
+import { html, LitElement, nothing } from 'lit';
 import { property, state } from 'lit/decorators.js';
+import { ifDefined } from 'lit/directives/if-defined.js';
 
-import '@warp-ds/icons/elements/check-16';
+import '../icon/icon.js';
 
-import { activateI18n } from '../i18n.js';
+import { activateI18n, detectLocale } from '../i18n.js';
+import { styles } from '../step-indicator/styles.js';
+import { reset } from '../styles.js';
 import { messages as daMessages } from './locales/da/messages.mjs';
 import { messages as enMessages } from './locales/en/messages.mjs';
 import { messages as fiMessages } from './locales/fi/messages.mjs';
 import { messages as nbMessages } from './locales/nb/messages.mjs';
 import { messages as svMessages } from './locales/sv/messages.mjs';
-import { styles } from '../step-indicator/styles.js';
-import { reset } from '../styles.js';
 
 const ccStep = {
   base: 'group/step',
@@ -39,24 +40,6 @@ const ccStep = {
   content: 'last:mb-0',
   contentVertical: 'row-span-2 pb-32',
   contentHorizontal: 'col-span-3 px-16 row-start-1 text-center',
-};
-
-const availableAriaLabels = {
-  completed: i18n._({
-    id: 'steps.aria.completed',
-    message: 'Step indicator completed circle',
-    comment: 'Completed circle',
-  }),
-  active: i18n._({
-    id: 'steps.aria.active',
-    message: 'Step indicator active circle',
-    comment: 'Active circle',
-  }),
-  default: i18n._({
-    id: 'steps.aria.emptyCircle',
-    message: 'Empty circle',
-    comment: 'Empty circle',
-  }),
 };
 
 export interface StepsContext {
@@ -93,13 +76,28 @@ export class WarpStep extends LitElement {
 
   setContext(context: StepsContext) {
     this._context = context;
-    this.requestUpdate();
   }
 
   getAriaLabel() {
-    if (this.completed) return availableAriaLabels.completed;
-    if (this.active) return availableAriaLabels.active;
-    return availableAriaLabels.default;
+    if (this.completed) {
+      return i18n._({
+        id: 'steps.aria.completed',
+        message: 'Step indicator completed circle',
+        comment: 'Completed circle',
+      });
+    }
+    if (this.active) {
+      return i18n._({
+        id: 'steps.aria.active',
+        message: 'Step indicator active circle',
+        comment: 'Active circle',
+      });
+    }
+    return i18n._({
+      id: 'steps.aria.emptyCircle',
+      message: 'Empty circle',
+      comment: 'Empty circle',
+    });
   }
 
   render() {
@@ -142,13 +140,13 @@ export class WarpStep extends LitElement {
     ]);
 
     return html`
-      <div class="${stepClasses}" style=${this._context.horizontal && 'height: 100%;'}>
-        ${!vertical ? html`<div class=${lineHorizontalClasses}></div>` : ''}
-        <div class=${dotClasses} role="img" aria-label=${this.getAriaLabel()} ${this.active ? 'aria-current="step"' : ''}>
-          ${this.completed ? html`<w-icon-check-16 data-testid="completed-icon"></w-icon-check-16>` : ''}
+      <div class="${stepClasses}" style=${ifDefined(this._context.horizontal ? 'height: 100%;' : undefined)}>
+        ${!vertical ? html`<div class=${lineHorizontalClasses}></div>` : nothing}
+        <div class=${dotClasses} role="img" aria-label=${this.getAriaLabel()} aria-current=${this.active ? 'step' : nothing}>
+          ${this.completed ? html`<w-icon name="Check" size="small" locale="${detectLocale()}" data-testid="completed-icon" class="flex"></w-icon>` : nothing}
         </div>
         <div class=${lineClasses}></div>
-        <div class=${contentClasses} ${this._context.horizontal && `style='height:100%;'`}>
+        <div class=${contentClasses} style=${ifDefined(this._context.horizontal ? 'height: 100%;' : undefined)}>
           <slot></slot>
         </div>
       </div>